Provide NLogEcsJsonContext.Default
