+++ src/Elastic.Ingest.Elasticsearch.CommonSchema/Elastic.Ingest.Elasticsearch.CommonSchema.csproj
@@ -10,7 +10,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Elastic.Ingest.Elasticsearch" Version="0.12.1" />
+    <PackageReference Include="Elastic.Ingest.Elasticsearch" Version="0.12.2" />
   </ItemGroup>
 
 </Project>
