+++ tests-integration/Elasticsearch.IntegrationDefaults/Elasticsearch.IntegrationDefaults.csproj
@@ -11,7 +11,7 @@
   <ItemGroup>
     <PackageReference Include="Elastic.Clients.Elasticsearch" Version="9.0.4" />
     <PackageReference Include="Elastic.Elasticsearch.Xunit" Version="0.6.0" />
-    <PackageReference Include="Elastic.Ingest.Elasticsearch" Version="0.12.1" />
+    <PackageReference Include="Elastic.Ingest.Elasticsearch" Version="0.12.2" />
 
   </ItemGroup>
